+++ src/pages/MainPage/components/index.ts
@@ -1,1 +1,3 @@
 export { default as FlowerFrame } from './FlowerFrame'
+export { default as MarryGoRound } from './MarryGoRound'
+export { default as CreatedBouquet } from './CreatedBouquet'
